Move <ListSection /> to /treePage
